Update the structure of files
