+++ src/ast.ts
@@ -1,4 +1,4 @@
-import * as commands from './commands.js';
+import * as commands from './stdlib/commands.js';
 import type * as types from './types.js';
 
 /** @internal */
@@ -7,9 +7,6 @@ export const enum _Message {
 	BLOCK_END_NOT_INIT = 'Block does not have a start offset',
 	START_IS_NOT_BLOCK_START = 'Start of block is not a BlockStart',
 }
-
-/** @public */
-export { NativeFunction } from './commands.js';
 
 /** @public */
 export class Literal implements types.ASTNode {
@@ -163,5 +160,41 @@ export class ImportStatement implements types.ASTNode {
 		const otherSource = context.getSource(resolved);
 
 		data.importSource(otherSource, `${this.namespace}:`);
+	}
+}
+
+/** @public */
+export class NativeFunction implements types.ASTNode {
+	constructor (
+		readonly name: string,
+		readonly executable: types.Executable,
+	) {}
+
+	static createArray (
+		map: ReadonlyArray<readonly [string, types.Executable]>,
+	): types.ASTTree {
+		const ast: types.ASTNode[] = map.map(([name, executable]) =>
+			new NativeFunction(name, executable));
+
+		return ast;
+	}
+
+	assemble ({ source, data, offset }: types.AssembleArg) {
+		data.addIdentifier(this.name, {
+			sourceName: source.name,
+			offset,
+			exported: true,
+			implicitlyCalled: true,
+		});
+	}
+
+	async execute (arg: types.ExecuteArg) {
+		const value = this.executable(arg);
+
+		if (value !== undefined) {
+			await value;
+		}
+
+		commands.return_(arg);
 	}
 }
