+++ test/stdlib/commands.d/frame.ts
@@ -1,32 +1,24 @@
 import { test } from 'tap';
-import { local_ } from '#src/commands.js';
+import { frame_ } from '#src/stdlib/commands.js';
 import { CommandsMessage } from '#test/test-util/message.js';
 import { createAssets } from '#test/test-util/stakr.js';
 
-await test('local', async (t) => {
+await test('frame', async (t) => {
 	const { data, arg } = await createAssets();
 	const array = [1, 2, 3];
 
 	data.stack.push(...array);
 	data.framePointer = 2;
-	local_(arg);
+	frame_(arg);
 
-	t.strictSame(data.stack.toNewArray(), [...array, data.framePointer],
+	t.strictSame(data.stack.toNewArray(), [...array, -1],
 		'expected to push to the stack');
-
-	data.stack.clear();
-	data.stack.push(...array);
-	data.framePointer = 0;
-	local_(arg);
-
-	t.strictSame(data.stack.toNewArray(), [...array, data.framePointer],
-		'expected to push to the stack even if frame pointer is zero');
 
 	data.framePointer = 0.3;
 
 	t.throws(
 		() => {
-			local_(arg);
+			frame_(arg);
 		},
 		new RangeError(CommandsMessage.FRAME_POINTER_IS_NOT_VALID),
 		'expected to throw if frame pointer is not an integer',
@@ -36,7 +28,7 @@ await test('local', async (t) => {
 
 	t.throws(
 		() => {
-			local_(arg);
+			frame_(arg);
 		},
 		new RangeError(CommandsMessage.FRAME_POINTER_IS_NOT_VALID),
 		'expected to throw if frame pointer is not a safe integer',
@@ -46,17 +38,27 @@ await test('local', async (t) => {
 
 	t.throws(
 		() => {
-			local_(arg);
+			frame_(arg);
 		},
 		new RangeError(CommandsMessage.FRAME_POINTER_IS_NOT_VALID),
 		'expected to throw if frame pointer is negative',
+	);
+
+	data.framePointer = 0;
+
+	t.throws(
+		() => {
+			frame_(arg);
+		},
+		new RangeError(CommandsMessage.FRAME_POINTER_IS_AT_START),
+		'expected to throw if frame pointer is zero',
 	);
 
 	data.framePointer = data.stack.length + 1;
 
 	t.throws(
 		() => {
-			local_(arg);
+			frame_(arg);
 		},
 		new RangeError(CommandsMessage.FRAME_POINTER_IS_PAST_END),
 		'expected to throw if frame pointer is more than stack length',
