+++ test/stdlib/commands.d/goto.ts
@@ -1,5 +1,5 @@
 import { test } from 'tap';
-import { goto_ } from '#src/commands.js';
+import { goto_ } from '#src/stdlib/commands.js';
 import testGoto from '#test/test-util/goto.js';
 import { createAssets } from '#test/test-util/stakr.js';
