+++ test/stdlib/commands.d/local.ts
@@ -1,24 +1,32 @@
 import { test } from 'tap';
-import { frame_ } from '#src/commands.js';
+import { local_ } from '#src/stdlib/commands.js';
 import { CommandsMessage } from '#test/test-util/message.js';
 import { createAssets } from '#test/test-util/stakr.js';
 
-await test('frame', async (t) => {
+await test('local', async (t) => {
 	const { data, arg } = await createAssets();
 	const array = [1, 2, 3];
 
 	data.stack.push(...array);
 	data.framePointer = 2;
-	frame_(arg);
+	local_(arg);
 
-	t.strictSame(data.stack.toNewArray(), [...array, -1],
+	t.strictSame(data.stack.toNewArray(), [...array, data.framePointer],
 		'expected to push to the stack');
+
+	data.stack.clear();
+	data.stack.push(...array);
+	data.framePointer = 0;
+	local_(arg);
+
+	t.strictSame(data.stack.toNewArray(), [...array, data.framePointer],
+		'expected to push to the stack even if frame pointer is zero');
 
 	data.framePointer = 0.3;
 
 	t.throws(
 		() => {
-			frame_(arg);
+			local_(arg);
 		},
 		new RangeError(CommandsMessage.FRAME_POINTER_IS_NOT_VALID),
 		'expected to throw if frame pointer is not an integer',
@@ -28,7 +36,7 @@ await test('frame', async (t) => {
 
 	t.throws(
 		() => {
-			frame_(arg);
+			local_(arg);
 		},
 		new RangeError(CommandsMessage.FRAME_POINTER_IS_NOT_VALID),
 		'expected to throw if frame pointer is not a safe integer',
@@ -38,27 +46,17 @@ await test('frame', async (t) => {
 
 	t.throws(
 		() => {
-			frame_(arg);
+			local_(arg);
 		},
 		new RangeError(CommandsMessage.FRAME_POINTER_IS_NOT_VALID),
 		'expected to throw if frame pointer is negative',
-	);
-
-	data.framePointer = 0;
-
-	t.throws(
-		() => {
-			frame_(arg);
-		},
-		new RangeError(CommandsMessage.FRAME_POINTER_IS_AT_START),
-		'expected to throw if frame pointer is zero',
 	);
 
 	data.framePointer = data.stack.length + 1;
 
 	t.throws(
 		() => {
-			frame_(arg);
+			local_(arg);
 		},
 		new RangeError(CommandsMessage.FRAME_POINTER_IS_PAST_END),
 		'expected to throw if frame pointer is more than stack length',
